+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.0 - In Development
 
 * [#61](https://github.com/adamcaudill/yawast/issues/61) - SSL Session Count: force 3DES suites
+* [#23](https://github.com/adamcaudill/yawast/issues/23) - Add check for HTTP to HTTPS redirect
 
 ## 0.2.2 - 2016-08-07